requirements.txt
streamlit
opencv-python-headless
numpy<2.0.0
Pillow
requests
